# apps/api/src/resources/swagger.yaml
...
openapi: 3.0.0
info:
  title: PCGL DACO API
  version: 1.0.0
paths:
...
  /signature/{applicationId}:
    get:
      tags:
        - Signature
      summary: Get the signatures associated with an application.
      parameters:
        - name: applicationId
          in: path
          description: The ID of the application you wish to retrieve the signature from.
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Successfully retrieved signature details from an application record. Note that these may be `null`, depending on if the application has been signed or not.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/SignatureRecord'
        '400':
          description: Will occur when not provided a valid `applicationId`, either a negative number, or not a number at all.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/responses/RequestValidationError'
                  - $ref: '#/components/responses/ClientErrors'
        '404':
          description: Application not found.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ClientErrors'
        '500':
          description: Something went wrong while processing your request.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerErrors'
    delete:
      tags:
        - Signature
      summary: Delete signatures associated with an application.
      parameters:
        - name: applicationId
          in: path
          description: The ID of the application you wish to retrieve the signature from.
          required: true
          schema:
            type: integer
            minimum: 1
        - in: query
          name: signee
          schema:
            type: string
            enum: [INSTITUTIONAL_REP, APPLICANT]
          description: Which signature to delete
          required: true
      responses:
        '204':
          description: Successfully deleted the signature requested. This will always result in a `No Content` response.
        '400':
          description: Will occur when not provided a valid `applicationId`, either a negative number, or not a number at all or invalid `signee`.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/responses/RequestValidationError'
                  - $ref: '#/components/responses/ClientErrors'
        '404':
          description: Application not found.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ClientErrors'
        '500':
          description: Something went wrong while processing your request.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerErrors'
  /signature/sign:
    post:
      tags:
        - Signature
      summary: Sign an application with a signature.
      requestBody:
        description: The ID of the application you wish to add the signature to, the signature (encoded in base64) and the time the signature took place.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/requests/SignatureSign'
      responses:
        '200':
          description: Application has been successfully signed. It will return the same data to you as what you sent in with a status `200 OK`.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/SignatureSigned'
        '400':
          description: Body parameters incorrect / validation failed. Likely due to missing params, or a non-base64 string.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/responses/RequestValidationError'
                  - $ref: '#/components/responses/ClientErrors'
        '404':
          description: Application not found.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ClientErrors'
        '500':
          description: Something went wrong while processing your request.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerErrors'
...
components:
...
  requests:
...
    SignatureSign:
      type: object
      description: The contents of the signature details in an application.
      properties:
        applicationId:
          type: number
          description: The ID of the application you want to add/modify the signature on.
        signature:
          pattern: '^(?:data\:image\/png\;base64,(\s+)?){1}([0-9a-zA-Z+/]{4})*(([0-9a-zA-Z+/]{2}==)|([0-9a-zA-Z+/]{3}=))?$'
          type: string
          description: |
            The Base64 encoded PNG image containing the signature of the Applicant or Institutional Representative.
        signee:
          type: string
          enum: [APPLICANT, INSTITUTIONAL_REP]
          description: Whether the signature is being performed by an applicant or a Institutional Representative.
...
  responses:
...
    ClientErrors:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: General description of error with parameters and context
...
    ServerErrors:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: General description of error with parameters and context
        errors:
          description: Detailed error data thrown by code failure
          type: string
    SignatureRecord:
      type: object
      description: The contents of the signature details contained within an application record.
      properties:
        applicationId:
          type: number
          description: The ID of the application you want to add/modify the signature on.
        applicantSignature:
          pattern: '^(?:data\:image\/png\;base64,(\s+)?){1}([0-9a-zA-Z+/]{4})*(([0-9a-zA-Z+/]{2}==)|([0-9a-zA-Z+/]{3}=))?$'
          type: string
          description: |
            The Base64 encoded PNG image containing the signature of the Applicant.
          nullable: true
        applicantSignedAt:
          type: string
          format: date-time
          description: The time the applicant signed at.
          nullable: true
        institutionalRepSignature:
          pattern: '^(?:data\:image\/png\;base64,(\s+)?){1}([0-9a-zA-Z+/]{4})*(([0-9a-zA-Z+/]{2}==)|([0-9a-zA-Z+/]{3}=))?$'
          type: string
          description: |
            The Base64 encoded PNG image containing the signature of the Applicant.
          nullable: true
        institutionalRepSignedAt:
          type: string
          format: date-time
          description: The time the applicant signed at.
          nullable: true
    SignatureSigned:
      type: object
      description: The contents of the signature details in an application.
      properties:
        id:
          type: number
          description: The ID of the application that you modified the signature on.
        signature:
          pattern: '^(?:data\:image\/png\;base64,(\s+)?){1}([0-9a-zA-Z+/]{4})*(([0-9a-zA-Z+/]{2}==)|([0-9a-zA-Z+/]{3}=))?$'
          type: string
          description: |
            The Base64 encoded PNG image containing the signature of the Applicant or Institutional Representative.
        signedAt:
          type: string
          format: date-time
          description: The the signature was signed.
    RequestValidationError:
      type: object
      required:
        - message
        - error
        - details
      properties:
        error:
          type: string
          description: Error code, usually "RequestValidationError"
        message:
          type: string
          description: General description of error with parameters and context
        details:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
                description: The validation error type
              expected:
                type: string
                description: The expected property.
              received:
                type: string
                description: The property that was received in the request
              path:
                type: array
                items:
                  type: string
                  description: The path of the object where validation failed.
              message:
                type: string
                description: A plain text error message describing why the validation failed.